requirements.txt
accelerate==0.26.1
agentops==0.3.26
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.20.0
airportsdata==20250224
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
apify_client==1.9.2
apify_shared==1.3.2
arrow==1.3.0
arxiv==2.1.3
arxiv2text==0.1.14
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
av==14.2.0
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
bibtexparser==1.4.3
botocore==1.37.23
cachetools==5.5.2
camel-ai==0.2.32
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
chunkr-ai==0.0.44
click==8.1.8
cloudpickle==3.1.1
cohere==5.14.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.7.1
cryptography==44.0.2
cssselect==1.3.0
cssutils==2.11.1
curl_cffi==0.6.2
currency-symbols==2.0.3
cycler==0.12.1
dappier==0.3.4
dataclasses-json==0.6.7
datacommons==1.4.3
datacommons-pandas==0.0.3
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.25.1
dill==0.3.8
discord.py==2.5.2
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring-parser==0.15
docutils==0.21.2
docx==0.2.4
docx2markdown==0.1.1
docx2txt==0.8
duckdb==1.2.1
duckduckgo_search==6.4.2
e2b==1.3.2
e2b-code-interpreter==1.2.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.2.0
fake-useragent==2.1.0
Faker==19.13.0
fastapi==0.115.12
fastavro==1.10.0
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
firecrawl-py==1.6.8
fish-audio-sdk==2024.12.5
fonttools==4.56.0
fpdf==1.7.2
free-proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
genson==1.3.0
geojson==2.5.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
googlemaps==4.10.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.29.3
hyperframe==6.1.0
identify==2.6.9
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.34.0
isodate==0.7.2
jaraco.context==6.0.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langdetect==1.0.9
lark==1.2.2
lazy-object-proxy==1.10.0
linkup-sdk==0.2.4
litellm==1.65.0
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.6.0
mcp-server-fetch==2025.1.17
mcp-simple-arxiv==0.2.2
mdurl==0.1.2
mem0ai==0.1.81
milvus-lite==2.4.12
mistralai==1.5.2
mock==5.2.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mypy==1.15.0
mypy-extensions==1.0.0
narwhals==1.32.0
nebula3-python==3.8.2
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
nodeenv==1.9.1
notion-client==2.3.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.47
openai==1.69.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
outcome==1.3.0.post0
outlines==0.1.14
outlines_core==0.1.26
-e git+https://github.com/camel-ai/owl.git@3641e4b04fba4e449fb982b1393e62328475b8d6#egg=owl
packaging==24.2
pandas==1.5.3
pandasai==2.4.2
parso==0.8.4
pathable==0.4.4
pdfminer.six==20250327
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.23.0
prance==23.6.21.0
praw==7.8.1
prawcore==2.4.0
pre-commit==3.8.0
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.1
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.8.1
pydantic_core==2.23.4
pydub==0.25.1
pyee==12.1.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pymilvus==2.5.6
PyMuPDF==1.25.4
PyMySQL==1.1.1
PyNaCl==1.5.0
pyowm==3.3.0
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
pytest==7.4.4
pytest-asyncio==0.23.8
pytest-cov==4.1.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytidb-experimental==0.0.1.dev4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
qdrant-client==1.13.3
rank-bm25==0.2.2
RapidFuzz==3.12.2
readabilipy==0.3.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
reka-api==3.2.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rouge==1.0.1
rpds-py==0.24.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scenedetect==0.6.6
scholarly==1.7.11
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
semantic-version==2.10.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
sqlglot==25.34.1
sqlglotrs==0.3.0
sqlmodel==0.0.22
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
stem==1.8.2
stripe==11.6.0
sympy==1.13.3
tabulate==0.9.0
tavily-python==0.5.3
termcolor==2.4.0
textblob==0.17.1
threadpoolctl==3.6.0
tidb-vector==0.0.14
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
tomli==2.2.1
torch==2.4.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
tree-sitter==0.23.2
tree-sitter-python==0.23.6
trio==0.29.0
trio-websocket==0.12.2
triton==3.0.0
typer==0.15.2
types-colorama==0.4.15.20240311
types-mock==5.2.0.20250306
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20250326
types-requests==2.32.0.20250328
types-setuptools==69.5.0.20240522
types-tqdm==4.67.0.20250319
typing-inspect==0.9.0
typing_extensions==4.13.0
ujson==5.10.0
unstructured==0.17.2
unstructured-client==0.27.0
update-checker==0.18.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
xls2xlsx==0.2.0
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2024.12.23
zipp==3.21.0
